refactor: move u5c conversion out of balius-proto into the runtime adapter
balius-proto is the Balius-owned wire schema crate; coupling it to an
arbitrary upstream service (utxorpc-spec) defeats the point of having a
schema crate in the first place. The conversion is u5c-specific and only
consumed by balius-runtime, so it moves into the u5c ledger adapter.

balius-proto/src/convert.rs is deleted along with its `convert` feature
and the `utxorpc-spec` / `serde_json` deps. The crate is now pure types
plus wire-format invariant assertions.

The TryFrom impls become plain `convert_tx` / `convert_tx_output` / etc.
functions inside `balius-runtime/src/ledgers/u5c.rs` (the orphan rule
forbids the impls here — neither side is local to balius-runtime).
`ConvertError` and `pparams_to_legacy_json` move alongside them. Call
sites in `lib.rs::Block::txs` and the ledger paths update from
`.try_into()` to the named functions. No behavior change.

Conversion tests move from `balius-proto/tests/wire_compat.rs` to a new
`balius-runtime/tests/u5c_convert.rs`. balius-proto keeps only the two
`wire_compat_*` tests — the wire-format invariant of the schema itself,
no upstream involved.

Verified: cargo check --workspace clean; 2/2 wire-compat tests in
balius-proto; 8/8 u5c convert tests in balius-runtime; e2e::faucet_claim
still passes (end-to-end ABI compat path through the mock ledger);
u5c-chainsync::wallet_balance still fails on the same pre-existing
stale-wasm 5-vs-7 WIT variant mismatch, unrelated.
