+++ balius-runtime/Cargo.toml
@@ -25,7 +25,7 @@ itertools = "0.14.0"
 async-trait = "0.1.83"
 utxorpc = { version = "0.13.0" }
 # utxorpc = { path = "../../../utxorpc/rust-sdk" }
-balius-proto = { version = "0.5.2", path = "../balius-proto", features = ["convert"] }
+balius-proto = { version = "0.5.2", path = "../balius-proto" }
 tokio-util = "0.7.12"
 prost = "0.13"
 object_store = { version = "0.12.0", features = ["fs"] }
@@ -39,6 +39,7 @@ futures = "0.3.31"
 tokio = "1.40.0"
 wat = "1"
 wit-component = "0.225"
+utxorpc-spec_017 = { package = "utxorpc-spec", version = "0.17.0", features = ["ledgers"], default-features = false }
 
 [features]
 http = ["object_store/http"]
